Split the python file into separate parts. Changed the structure of the files.
